+++ examples/src/unamed.ml
@@ -1,0 +1,3 @@
+type (_, 'b, 'a) t = int
+
+let foo _ _ _ = true