+++ examples/src/unamed.mli
@@ -1,0 +1,3 @@
+type (_, 'b, 'a) t
+
+val foo : 'a -> _ -> 'b -> bool