+++ src/compat.ml
@@ -1,2 +1,3 @@
 let concat_map = List.concat_map
 let find_map = List.find_map
+let fold_left_map = List.fold_left_map